test/wpt/jsdom/issues/3055.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>3055</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<!-- Regression test for https://github.com/jsdom/jsdom/issues/3055 -->

<div class="container" id="container">
  Focus here:
  <button id="item">focus me</button>
</div>

<script>
"use strict";
const container = document.getElementById('container');
const item = document.getElementById('item');
test(() => {
  item.focus();
  assert_true(item.matches(':focus-within'));
  item.blur();
}, 'item matches :focus-within');
test(() => {
  item.focus();
  assert_true(container.matches(':focus-within'));
  item.blur();
}, 'parent matches :focus-within');
test(() => {
  item.focus();
  assert_true(document.body.matches(':focus-within'));
  item.blur();
}, 'ancestor matches :focus-within');
</script>
